Add categories and time slider to arc deck.gl viz (#5638)
* Fix legend position

* Add categories and play slider to arc viz

* New functionality to arc viz

(cherry picked from commit 6959b70c1c2801b0c828b332fa79f562495fcd2f)
